IK-91 [Refactored] Moved User.java to userDomain package.
