+++ client/app/src/main/java/com/example/mrides/RequestQueueSingleton.java
@@ -34,9 +34,10 @@ public class RequestQueueSingleton {
 
             mInstance = new RequestQueueSingleton(context);
         }
-
+        mCtx = context;
         return mInstance;
     }
+
 
     public RequestQueue getRequestQueue() {
