+++ client/app/src/main/java/com/example/mrides/SignInRequestHandler.java
@@ -10,13 +10,8 @@ import android.content.Context;
 import com.android.volley.Request;
 import com.android.volley.Response;
 import com.android.volley.VolleyError;
-import com.android.volley.toolbox.JsonObjectRequest;
 import com.android.volley.toolbox.StringRequest;
-import com.example.mrides.Domain.User;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 
-import org.json.JSONObject;
 import java.util.HashMap;
 import java.util.Map;
